Move parsing into separate submodule location
